make it better

+++ ePark/src/main/java/com/ePark/controller/ErrorsController.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.servlet.error.ErrorController;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.ePark.service.BookingFlow;
@@ -19,38 +20,67 @@ import com.ePark.service.BookingService;
 import ch.qos.logback.classic.Logger;
 
 @Controller
-public class ErrorsController implements ErrorController{
-	
+public class ErrorsController implements ErrorController {
+
 	@Autowired
 	private BookingService bookingService;
-	
+
 	private static final String PATH = "/error";
 
-    @RequestMapping(value = PATH)
-    @Transactional
-    public String handleError(HttpServletRequest httpRequest) {
-    	Object status = httpRequest.getAttribute(RequestDispatcher.ERROR_STATUS_CODE);
-    	Integer statusCode = Integer.valueOf(status.toString());
-    	
-        if(statusCode == HttpStatus.NOT_FOUND.value()) {
-        	Logger logger = (Logger) LoggerFactory.getLogger(ErrorsController.class);
-            logger.error("An exception occurred!");
-        }
-        
-        
-    	HttpSession session = httpRequest.getSession();
-    	BookingFlow bookingFlow = (BookingFlow) session.getAttribute("bookingFlow");
-    
-    	if (bookingFlow != null) {
-    		bookingService.delete(bookingFlow.getBooking().getBookingId());
-    	}
-        
-        return "error";
-    }
+	@RequestMapping(value = PATH)
+	@Transactional
+	public String handleError(HttpServletRequest httpRequest, Model model) {
+		/*
+		 * Object status =
+		 * httpRequest.getAttribute(RequestDispatcher.ERROR_STATUS_CODE); Integer
+		 * statusCode = Integer.valueOf(status.toString());
+		 * 
+		 * if(statusCode == HttpStatus.NOT_FOUND.value()) { Logger logger = (Logger)
+		 * LoggerFactory.getLogger(ErrorsController.class);
+		 * logger.error("An exception occurred!"); }
+		 */
 
-    @Override
-    public String getErrorPath() {
+		HttpSession session = httpRequest.getSession();
+		BookingFlow bookingFlow = (BookingFlow) session.getAttribute("bookingFlow");
 
-        return PATH;
-    }
+		if (bookingFlow != null) {
+			if (bookingFlow.getBooking().getBookingId() != 0) {
+				bookingService.delete(bookingFlow.getBooking().getBookingId());
+			}
+
+		}
+
+		
+		
+	    Object status = httpRequest.getAttribute(RequestDispatcher.ERROR_STATUS_CODE);
+	    String message = null;
+	    
+	    if (status != null) {
+	        Integer statusCode = Integer.valueOf(status.toString());
+	    
+	        if(statusCode == HttpStatus.NOT_FOUND.value()) {
+	        	
+	        	message = "Page not found";
+	        	
+	        } else if(statusCode == HttpStatus.INTERNAL_SERVER_ERROR.value()) {
+	        	
+	            message = "Internal server error";
+	            
+	        } else if(statusCode == HttpStatus.FORBIDDEN.value()) {
+	        	
+	            message = "You do not have permission to access this page";
+	        }
+	    }
+		
+
+		model.addAttribute("message", message);
+
+		return "error";
+	}
+
+	@Override
+	public String getErrorPath() {
+
+		return PATH;
+	}
 }
